[IN-21] refactor profile image to component
