+++ src/components/profile-image/index.ts
@@ -1,0 +1,1 @@
+export { ProfileImage } from './profile-image'
