+++ src/components/profile-image/profile-image.module.scss
@@ -1,0 +1,94 @@
+.userImageBox {
+  position: relative;
+
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  align-items: center;
+
+  padding-top: 48px;
+}
+
+.userImage {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  width: 198px;
+  height: 198px;
+
+  background-color: var(--dark-500);
+  border-radius: 50%;
+}
+
+.addUserImage {
+  position: relative;
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  width: 332px;
+  height: 340px;
+  margin-bottom: 60px;
+
+  background-color: var(--dark-500);
+}
+
+.line2 {
+  position: absolute;
+  right: 0;
+  bottom: 36px;
+  left: 0;
+
+  width: 120%;
+  height: 1px;
+  margin-bottom: 24px;
+
+  background-color: var(--dark-300);
+}
+
+.modalContent {
+  width: 492px;
+  min-height: 564px;
+}
+
+.modalInnerContent {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  padding-top: 48px;
+  padding-bottom: 84px;
+}
+
+.modalInnerContent:has(.saveButton) {
+  padding-top: 4px;
+  padding-bottom: 12px;
+}
+
+.saveButton {
+  align-self: flex-end;
+}
+
+.profileAvatar {
+  position: relative;
+
+  .deleteAvatarIcon {
+    position: absolute;
+    top: 0;
+    right: 0;
+
+    width: 24px;
+    height: 24px;
+    padding: 0;
+
+    background-color: transparent;
+  }
+}
+
+.uploadError {
+  position: absolute;
+  bottom: -30px;
+  color: var(--danger-500);
+}
